extract sensei-core, move integration test
