+++ senseicore/Cargo.toml
@@ -1,18 +1,12 @@
 [package]
-name = "sensei"
-version = "0.2.1"
+name = "senseicore"
+version = "0.1.0"
 edition = "2021"
+publish = false
 
-[workspace]
-members = [".", "entity", "migration", "senseicore"]
-
-[[bin]] # Bin to run sensei cli
-name = "senseicli"
-path = "src/cli.rs"
-
-[[bin]] # Bin to run sensei daemon
-name = "senseid"
-path = "src/main.rs"
+[lib]
+name = "senseicore"
+path = "src/lib.rs"
 
 [dependencies]
 lightning = { version = "0.0.106", features = ["max_level_trace"], git = "https://github.com/lightningdevkit/rust-lightning", rev = "d0f69f77bd6ed40bff7ef1026f23e4444a5a884a" }
@@ -38,33 +32,20 @@ log = "0.4.16"
 env_logger = "0.9.0"
 bitcoincore-rpc = "0.14"
 bdk = { version = "0.18.0" }
-tonic = "0.6"
-prost = "0.9"
 pin-project = "1.0"
 hyper = "0.14"
 clap = { version = "3.0", features = [ "derive", "env" ] }
 headers = "0.3"
 tindercrypt = { version = "0.3.2", default-features = false }
 portpicker = "0.1"
-rusqlite = { version = "0.25.3", features = ["uuid", "bundled"] }
 uuid = { version = "0.8", features = ["serde", "v4"] }
 macaroon = "0.2"
-tower-http = { version = "0.2.5", features = ["fs", "trace", "cors"] }
-tower-cookies = "0.4"
 dirs = "4.0"
 public-ip = "0.2"
 rust-embed="6.3.0"
 mime_guess = { version = "2" }
-entity = { path = "entity" }
-migration = { path = "migration" }
-senseicore = { path = "senseicore" }
-
-[build-dependencies]
-tonic-build = "0.6"
+entity = { path = "../entity" }
+migration = { path = "../migration" }
 
 [dev-dependencies]
 bitcoind = { version = "0.25.0", features = [ "22_0" ] }
-
-[[test]]
-name = "senseicore"
-path = "senseicore/tests/tests.rs"
