+++ senseicore/src/disk.rs
