+++ senseicore/src/error.rs
