+++ senseicore/src/hex_utils.rs
