+++ senseicore/src/node.rs
@@ -12,13 +12,13 @@ use crate::chain::database::WalletDatabase;
 use crate::chain::fee_estimator::SenseiFeeEstimator;
 use crate::chain::manager::SenseiChainManager;
 use crate::config::SenseiConfig;
+use crate::database::SenseiDatabase;
 use crate::disk::FilesystemLogger;
 use crate::error::Error;
 use crate::event_handler::LightningNodeEventHandler;
-use crate::lib::database::SenseiDatabase;
-use crate::lib::events::SenseiEvent;
-use crate::lib::network_graph::OptionalNetworkGraphMsgHandler;
-use crate::lib::persist::{AnyKVStore, DatabaseStore, SenseiPersister};
+use crate::events::SenseiEvent;
+use crate::network_graph::OptionalNetworkGraphMsgHandler;
+use crate::persist::{AnyKVStore, DatabaseStore, SenseiPersister};
 use crate::services::node::{Channel, NodeInfo, NodeRequest, NodeRequestError, NodeResponse, Peer};
 use crate::services::{PaginationRequest, PaginationResponse, PaymentsFilter};
 use crate::utils::PagedVec;
