+++ senseicore/src/persist.rs
