+++ senseicore/src/utils.rs
