+++ senseicore/src/version.rs
