+++ senseicore/tests/tests.rs
@@ -1,0 +1,1 @@
+pub mod smoke_test;
